added util file to clean up
