+++ hub/auto/directory_parsers.py
@@ -1,79 +1,18 @@
 import os
 from tqdm import tqdm
 import numpy as np
-from glob import glob
-
 from PIL import Image
 
 import hub
 
-__all__ = ['get_parsers']
-
-_parsers = []
-_priorities = []
+from hub.auto import util
 
 USE_TQDM = True
-IMAGE_EXTS = ['.jpg', '.png', '.jpeg']
 
 
-def get_image_shape(path):
-    img = Image.open(path)
-    c = len(img.getbands())
-    w, h = img.size
-    return (h, w, c)
-
-
-def get_children(path):
-    """helper function to glob the given directory"""
-
-    return glob(os.path.join(path, '*'))
-
-
-def get_ext(path):
-    ext = os.path.splitext(path)[-1]
-    return ext.lower()
-
-
-def files_are_of_extension(path, allowed_extensions):
-    """helper function that checks if all files within the given directory have extensions that are allowed"""
-
-    allowed_extensions = [ext.lower() for ext in allowed_extensions]
-    children = get_children(path)
-    # print(set([get_ext(child) for child in children]))
-    # print(set([child for child in children if get_ext(child) == '']))
-    return any([get_ext(child) in allowed_extensions for child in children])
-
-
-def get_parsers(priority_sort=True):
-    if priority_sort:
-        sorted_parsers = [x for _, x in sorted(zip(_priorities, _parsers))]
-        return sorted_parsers
-    return _parsers
-
-
-def directory_parser(priority=0):
-    """
-    a directory parser function is a function that takes in a path & returns a dataset.
-    these functions make it easier to extend the dataset inference domain. 
-    functions should be as general as possible.
-
-    Parameters
-    ----------
-    priority: int
-        an arbitrary number that the parsers will be sorted by
-        (lower the number = higher the priority)
-    """
-    def decorate(fn):
-        _parsers.append(fn)
-        _priorities.append(priority)
-        return fn
-
-    return decorate
-
-
-@directory_parser(priority=0)
+@util.directory_parser(priority=0)
 def image_classification(path, scheduler='single', workers=1):
-    children = get_children(path)
+    children = util.get_children(path)
 
     # check if all children are directories
     if not all([os.path.isdir(child) for child in children]):
@@ -81,7 +20,7 @@ def image_classification(path, scheduler='single', workers=1):
 
     # check if children's contents are all image files
     for child in children:
-        if not files_are_of_extension(child, IMAGE_EXTS):
+        if not util.files_are_of_extension(child, util.IMAGE_EXTS):
             return None
 
     # parse dataset
@@ -90,16 +29,17 @@ def image_classification(path, scheduler='single', workers=1):
     max_shape = np.zeros(3, dtype=np.uint8)  # CHW
     for child in tqdm(children,
                       desc='parsing image classification dataset',
-                      total=len(children)):
+                      total=len(children),
+                      disable=not USE_TQDM):
         label = os.path.basename(child)
 
-        filepaths = get_children(child)
+        filepaths = util.get_children(child)
         for filepath in filepaths:
             # ignore non-image extension files
-            if get_ext(filepath) not in IMAGE_EXTS:
+            if util.get_ext(filepath) not in util.IMAGE_EXTS:
                 continue
 
-            shape = np.array(get_image_shape(filepath))
+            shape = np.array(util.get_image_shape(filepath))
             max_shape = np.maximum(max_shape, shape)
             data.append((filepath, label.lower()))
             class_names.add(label.lower())
